GUACAMOLE-641: Rename guacamole-auth-vault to guacamole-vault (does not provide auth).
