--- extensions/guacamole-auth-vault/modules/guacamole-auth-vault-azure/src/main/resources/guac-manifest.json
@@ -1,16 +1,0 @@
-{
-
-    "guacamoleVersion" : "1.4.0",
-
-    "name"      : "Azure Key Vault",
-    "namespace" : "azure-keyvault",
-
-    "authProviders" : [
-        "org.apache.guacamole.auth.vault.azure.AzureKeyVaultAuthenticationProvider"
-    ],
-
-    "translations" : [
-        "translations/en.json"
-    ]
-
-}
