--- extensions/guacamole-auth-vault/modules/guacamole-auth-vault-dist/pom.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    Licensed to the Apache Software Foundation (ASF) under one
-    or more contributor license agreements.  See the NOTICE file
-    distributed with this work for additional information
-    regarding copyright ownership.  The ASF licenses this file
-    to you under the Apache License, Version 2.0 (the
-    "License"); you may not use this file except in compliance
-    with the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing,
-    software distributed under the License is distributed on an
-    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-    KIND, either express or implied.  See the License for the
-    specific language governing permissions and limitations
-    under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
-                        http://maven.apache.org/maven-v4_0_0.xsd">
-
-    <modelVersion>4.0.0</modelVersion>
-    <groupId>org.apache.guacamole</groupId>
-    <artifactId>guacamole-auth-vault-dist</artifactId>
-    <packaging>pom</packaging>
-    <name>guacamole-auth-vault-dist</name>
-    <url>http://guacamole.apache.org/</url>
-
-    <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
-
-    <parent>
-        <groupId>org.apache.guacamole</groupId>
-        <artifactId>guacamole-auth-vault</artifactId>
-        <version>1.4.0</version>
-        <relativePath>../../</relativePath>
-    </parent>
-
-    <dependencies>
-
-        <!-- Azure Key Vault Extension -->
-        <dependency>
-            <groupId>org.apache.guacamole</groupId>
-            <artifactId>guacamole-auth-vault-azure</artifactId>
-            <version>1.4.0</version>
-        </dependency>
-
-   </dependencies>
-
-    <build>
-
-        <!-- Dist .tar.gz for guacamole-auth-vault should be named after the
-            parent guacamole-auth-vault project, not after
-            guacamole-auth-vault-dist -->
-        <finalName>${project.parent.artifactId}-${project.parent.version}</finalName>
-
-    </build>
-
-</project>
