--- extensions/guacamole-auth-vault/modules/guacamole-auth-vault-dist/src/main/assembly/dist.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    Licensed to the Apache Software Foundation (ASF) under one
-    or more contributor license agreements.  See the NOTICE file
-    distributed with this work for additional information
-    regarding copyright ownership.  The ASF licenses this file
-    to you under the Apache License, Version 2.0 (the
-    "License"); you may not use this file except in compliance
-    with the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing,
-    software distributed under the License is distributed on an
-    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-    KIND, either express or implied.  See the License for the
-    specific language governing permissions and limitations
-    under the License.
--->
-<assembly
-    xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0 http://maven.apache.org/xsd/assembly-1.1.0.xsd">
-
-    <id>dist</id>
-    <baseDirectory>${project.parent.artifactId}-${project.parent.version}</baseDirectory>
-
-    <!-- Output .tar.gz -->
-    <formats>
-        <format>tar.gz</format>
-    </formats>
-
-    <!-- Include extension .jars -->
-    <dependencySets>
-
-        <!-- Azure Key Vault extension .jar -->
-        <dependencySet>
-            <outputDirectory>azure</outputDirectory>
-            <includes>
-                <include>org.apache.guacamole:guacamole-auth-vault-azure</include>
-            </includes>
-        </dependencySet>
-
-    </dependencySets>
-
-    <!-- Licenses -->
-    <fileSets>
-        <fileSet>
-            <outputDirectory></outputDirectory>
-            <directory>target/licenses</directory>
-        </fileSet>
-    </fileSets>
-
-</assembly>
