+++ extensions/guacamole-vault/modules/guacamole-vault-base/src/main/resources/translations/en.json
